requirements.txt
pyserial==3.4
kafka-python==1.4.4
jaeger-client==3.10.0
speedtest-cli==2.0.2